allmost works on loadmode, need to change fruit spawn at the start of the game

+++ pacman/Entity.cpp
@@ -146,7 +146,7 @@ void Entity::resetEntity() {
 	setY(defaultY);
 }
 
-short Entity::getLastMove() {
+int Entity::getLastMove() {
 	return lastmove;
 }
